Refactor: Updated the error response object structure and module exports
